Restructure Postgres workshop
Use shell scripts to automate DB setup for ease
