+++ src/workshops/node-postgres/starter-files/solution/example.env
@@ -1,0 +1,1 @@
+DATABASE_URL="postgres://learn_pg_user:pass123@localhost:5432/learn_pg"
