+++ src/workshops/node-postgres/starter-files/solution/package.json
@@ -1,21 +1,17 @@
 {
   "name": "node-postgres",
-  "version": "1.0.0",
-  "description": "",
+  "private": true,
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "dev": "nodemon server.js"
+    "dev": "nodemon -r dotenv/config server.js"
   },
-  "keywords": [],
-  "author": "Oliver Phillips <hello@oliverjam.es> (https://oliverjam.es/)",
-  "license": "MIT",
   "dependencies": {
-    "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "pg": "^8.6.0"
+    "pg": "^8.7.1"
   },
   "devDependencies": {
-    "nodemon": "^2.0.7"
+    "dotenv": "^10.0.0",
+    "nodemon": "^2.0.13"
   }
 }
