+++ src/workshops/node-postgres/starter-files/solution/routes/home.js
@@ -1,7 +1,7 @@
 const db = require("../database/connection.js");
 
 function get(request, response) {
-  db.query("SELECT * FROM users").then((result) => {
+  db.query("SELECT username FROM users").then((result) => {
     const users = result.rows;
     const userList = users.map((user) => `<li>${user.username}</li>`).join("");
     response.send(`<ul>${userList}</ul>`);
