+++ src/workshops/node-postgres/starter-files/solution/routes/posts.js
@@ -1,24 +1,23 @@
 const db = require("../database/connection.js");
 
 function get(request, response) {
-  const SELECT_POSTS = `
+  const select_posts = /*sql*/ `
     SELECT blog_posts.text_content, users.username
     FROM blog_posts INNER JOIN users
     ON blog_posts.user_id = users.id
     ORDER BY blog_posts.id DESC
   `;
-  db.query(SELECT_POSTS).then((result) => {
+  db.query(select_posts).then((result) => {
     const posts = result.rows;
-    let postItems = "";
-    for (let post of posts) {
-      postItems += `
+    const postItems = posts.map((post) => {
+      return /*html*/ `
         <li>
          <p>${post.text_content}</p>
          <p>${post.username}</p>
         </li>
       `;
-    }
-    response.send(`<ul>${postItems}</ul>`);
+    });
+    response.send(`<ul>${postItems.join("")}</ul>`);
   });
 }
