+++ src/workshops/node-postgres/starter-files/solution/scripts/populate_db
@@ -1,0 +1,7 @@
+#! /bin/sh
+
+# stop script when an error occurs
+set -e
+
+psql learn_pg -q -f "./database/init.sql"
+echo "Populated database tables"
