+++ src/workshops/node-postgres/starter-files/workshop/database/init.sql
