+++ src/workshops/node-postgres/starter-files/workshop/routes/create-user.js
@@ -1,7 +1,5 @@
-const db = require("../database/connection.js");
-
 function get(request, response) {
-  response.send(`
+  response.send(/*html*/ `
     <form action="create-user" method="POST">
       <p>
         <label for="username">Username</label>
@@ -12,10 +10,6 @@ function get(request, response) {
         <input id="age" name="age" type="number">
       </p>
       <p>
-        <label for="location">Location</label>
-        <input id="location" name="location">
-      </p>
-      <p>
         <button type="submit">Create user</button>
       </p>
     </form>
@@ -23,14 +17,8 @@ function get(request, response) {
 }
 
 function post(request, response) {
-  const data = request.body;
-  const values = Object.values(data);
-  db.query(
-    "INSERT INTO users(username, age, location) VALUES($1, $2, $3)",
-    values
-  ).then(() => {
-    response.redirect("/");
-  });
+  console.log(request.body); // e.g. { username: "oli", ... }
+  response.redirect("/");
 }
 
 module.exports = { get, post };
